requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
google-api-python-client
google-auth
google-auth-oauthlib